+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Unreleased changes. Release notes have not yet been written.
 * adds `DbInt` and `DbString` to key types.
 * removes `key_cache` and `record_cache` from features.
 * removes `kc-lfu` and `kc-lru` from features.
-* adds `get_k8()`, `put_k8()` and `del_k8()`.
+* adds `get_kt()`, `put_kt()` and `del_kt()`.
 * adds `htx-file` for supporting hash table index cache.
 * change separates dat-file to key-file and val-file.